ON-29826 # Added forms read claim. Updated arrays to lists for consistency
